Fixed error in simconnect data transmission. Moved simconnect dependency to handler down to cpp file. Added additional fields airplane category, num engines and engine type.

+++ src/fs/sc/simconnectreply.cpp
@@ -98,7 +98,7 @@ int SimConnectReply::write(QIODevice *ioDevice)
   // Go back and update size
   out.device()->seek(sizeof(MAGIC_NUMBER_REPLY));
   int size = block.size() - static_cast<int>(sizeof(packetSize)) - static_cast<int>(sizeof(MAGIC_NUMBER_REPLY));
-  out << static_cast<quint16>(size);
+  out << static_cast<quint32>(size);
 
   return SimConnectDataBase::writeBlock(ioDevice, block, status);
 }
